analytics/requirements.txt
jupyterlab
ipykernel
pandas
numpy
scipy
statsmodels
scikit-learn
matplotlib
seaborn
python-dotenv
supabase
